update documentation
yes, the drop from 51% documented to 41% documented is sort of legit.
These commits add a bunch of forwarding functions, and I moved all the
documentation to those functions, and removed it from the functions they
are being forwarded to, because I figured documentation for the public
interface should be in the public interface (the class).