simple waiting queue for requests
this should fix a lot of issues, as well as make it a ton easier to build
systems that won't blow up your api quotas

right now, it's just a simple "wait" queue, it needs to be implemented
in such a way so that there's a "waiting" and a "running" list, because
right now it's a hybrid of the old brute-force "re-run everything as soon
as a timer expires", and a better queue implementation.

throttlecheck has been updated to stagger it's requests, and uses
getMarketplaces rather than getProductCategories.